refactor: require explicit protocol params and add protocol assertions
Remove default values for hmssEnabled/trusTeeEnabled in all abstract
test classes so concrete classes must specify them explicitly.

Move cross-publisher creative-id test from DirectOnly to MultiEdp and
Gaussian noise failure test from concrete classes to abstract MultiEdp.

Add protocol assertions to all tests verifying the expected protocol
(direct, HMSS, or TrusTee) was used for the measurement.
