+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorDirectOnlyReportTest.kt
@@ -40,6 +40,8 @@ class GCloudEdpAggregatorDirectOnlyReportTest :
         reportingPostgresDatabaseProvider,
         IMPRESSION_QUALIFICATION_FILTER_MAPPING,
       ),
+    hmssEnabled = true,
+    trusTeeEnabled = true,
   ) {
 
   /**
