+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorHmssReportTest.kt
@@ -14,10 +14,8 @@
 
 package org.wfanet.measurement.integration.common.reporting.v2
 
-import kotlinx.coroutines.runBlocking
 import org.junit.ClassRule
 import org.junit.Rule
-import org.junit.Test
 import org.junit.rules.Timeout
 import org.wfanet.measurement.common.db.r2dbc.postgres.testing.PostgresDatabaseProviderRule
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
@@ -42,6 +40,7 @@ class GCloudEdpAggregatorHmssReportTest :
         reportingPostgresDatabaseProvider,
         IMPRESSION_QUALIFICATION_FILTER_MAPPING,
       ),
+    hmssEnabled = true,
     trusTeeEnabled = false,
   ) {
 
@@ -51,11 +50,6 @@ class GCloudEdpAggregatorHmssReportTest :
    * TODO(Kotlin/kotlinx.coroutines#3865): Switch back to CoroutinesTimeout when fixed.
    */
   @get:Rule val timeout: Timeout = Timeout.seconds(180)
-
-  @Test
-  fun `no noise basic report fails when EDP requires Gaussian noise`() = runBlocking {
-    assertReportFailsWhenEdpRequiresGaussianNoise()
-  }
 
   companion object {
     @get:ClassRule @JvmStatic val spannerEmulator = SpannerEmulatorRule()
