+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorLifeOfAnEventGroupTest.kt
@@ -40,6 +40,8 @@ class GCloudEdpAggregatorLifeOfAnEventGroupTest :
         reportingPostgresDatabaseProvider,
         IMPRESSION_QUALIFICATION_FILTER_MAPPING,
       ),
+    hmssEnabled = true,
+    trusTeeEnabled = true,
   ) {
 
   /**
