+++ src/test/kotlin/org/wfanet/measurement/integration/deploy/gcloud/GCloudInProcessHmssMultiEdpReportIntegrationTest.kt
@@ -37,6 +37,7 @@ class GCloudInProcessHmssMultiEdpReportIntegrationTest :
       reportingPostgresDatabaseProvider,
       IMPRESSION_QUALIFICATION_FILTER_MAPPING,
     ),
+    hmssEnabled = true,
     trusTeeEnabled = false,
   ) {
   companion object {
